requirements.txt
python-fasthtml>=0.13.0
python-dotenv>=1.0.0
monsterui>=0.1.0
resend>=0.3.0
